Small code style fix and new url to download foils data

+++ src/SetBulParameter.cpp
@@ -48,7 +48,7 @@ void SetBulbParam::ResetValues(void) {
 
 }
 
-float SetBulbParam::GetTargetWeight(void) {
+double SetBulbParam::GetTargetWeight(void) {
 
     return(ui->DSP_TW->value());
